watch_controller, argument improvements
`watch_controller()` was previously nod threaded, causing it to block the watching of a controller that was subsequently added. Now the method is threaded, and all controllers are watched simultaneously.

A global map `device_connected` indicates if a device is still connected. `main` updates the values in this map according to the udev monitor. `watch_controller()` checks this map with every iteration of its while loop. If the device is no longer connected, `watch_controller` returns.

The program now accepts a single argument: the timeout duration in seconds. This argument is options; the default value is 60.